image_caption_app/requirements.txt
# requirements.txt
streamlit>=1.36.0
openai>=1.48.0
requests>=2.31.0
pillow>=10.3.0
python-dotenv>=1.0.1
azure-identity>=1.17.1